# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrowserMind API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Browser agent task management
  - name: templates
    description: Task template management
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks with optional filters
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task details with steps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/execute:
    post:
      operationId: executeTask
      tags: [tasks]
      summary: Start task execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/pause:
    patch:
      operationId: pauseTask
      tags: [tasks]
      summary: Pause a running task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/resume:
    patch:
      operationId: resumeTask
      tags: [tasks]
      summary: Resume a paused task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{id}/stop:
    patch:
      operationId: stopTask
      tags: [tasks]
      summary: Stop/cancel a running task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/recent:
    get:
      operationId: getRecentTasks
      tags: [tasks]
      summary: Get the 5 most recent tasks
      responses:
        "200":
          description: Recent tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all saved task templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Save a task as a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"

  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /templates/{id}/execute:
    post:
      operationId: executeTemplate
      tags: [templates]
      summary: Create and execute a task from a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Task created from template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get dashboard statistics summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/settings:
    get:
      operationId: getSettings
      tags: [stats]
      summary: Get app settings/config
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    patch:
      operationId: updateSettings
      tags: [stats]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    TaskStep:
      type: object
      required: [id, taskId, stepNumber, action, status, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        stepNumber:
          type: integer
        action:
          type: string
        target:
          type: ["string", "null"]
        input:
          type: ["string", "null"]
        status:
          type: string
        screenshotUrl:
          type: ["string", "null"]
        reasoning:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    Task:
      type: object
      required: [id, title, instruction, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        instruction:
          type: string
        targetUrl:
          type: ["string", "null"]
        status:
          type: string
        outputFormat:
          type: string
        maxRetries:
          type: integer
        timeoutSeconds:
          type: integer
        executionTime:
          type: ["integer", "null"]
        stepCount:
          type: integer
        extractedData:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time

    TaskDetail:
      type: object
      required: [id, title, instruction, status, createdAt, steps]
      properties:
        id:
          type: integer
        title:
          type: string
        instruction:
          type: string
        targetUrl:
          type: ["string", "null"]
        status:
          type: string
        outputFormat:
          type: string
        maxRetries:
          type: integer
        timeoutSeconds:
          type: integer
        executionTime:
          type: ["integer", "null"]
        stepCount:
          type: integer
        extractedData:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TaskStep"

    TaskInput:
      type: object
      required: [instruction]
      properties:
        instruction:
          type: string
          minLength: 1
        title:
          type: string
        targetUrl:
          type: string
        outputFormat:
          type: string
        maxRetries:
          type: integer
        timeoutSeconds:
          type: integer

    Template:
      type: object
      required: [id, name, instruction, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        instruction:
          type: string
        targetUrl:
          type: ["string", "null"]
        outputFormat:
          type: string
        timesUsed:
          type: integer
        createdAt:
          type: string
          format: date-time

    TemplateInput:
      type: object
      required: [name, instruction]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        instruction:
          type: string
          minLength: 1
        targetUrl:
          type: string
        outputFormat:
          type: string

    StatsSummary:
      type: object
      required: [totalTasks, completedTasks, failedTasks, successRate, avgExecutionTime, activeTasksCount]
      properties:
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        failedTasks:
          type: integer
        successRate:
          type: number
        avgExecutionTime:
          type: number
        activeTasksCount:
          type: integer

    AppSettings:
      type: object
      required: [id, anthropicApiKey, defaultTimeout, autoRetry, screenshotFrequency, defaultOutputFormat, createdAt]
      properties:
        id:
          type: integer
        anthropicApiKey:
          type: ["string", "null"]
        defaultTimeout:
          type: integer
        autoRetry:
          type: boolean
        screenshotFrequency:
          type: string
        defaultOutputFormat:
          type: string
        createdAt:
          type: string
          format: date-time

    AppSettingsUpdate:
      type: object
      properties:
        anthropicApiKey:
          type: string
        defaultTimeout:
          type: integer
        autoRetry:
          type: boolean
        screenshotFrequency:
          type: string
        defaultOutputFormat:
          type: string